Fixed the form for adding a node
Now, when the user choses to add a node, the initial form asking them which function they wish to perform becomes hidden and the form asking them which value to they wish to insert is made visible. After the user hits the "insert" button the form asking them which value to insert becomes hidden and the initial form is made visible again. This process is repeatable.