Added support to remove node from webpage
Added necessary code to remove a value from the list via user input in webpage form. Much like for inserting a value, if the user chooses to remove a node the initial form is hidden and the form to remove a value is made visible. After the user hits the "remove" button, this remove form becomes hidden and the initial form is made visible again.