Implemented printing list from webpage form
User can now give input to print the list from the webpage form. The printed list is still, however, displayed via an alert() pop up. The implementation of this functionality is vey similar to the other functions, with the only main difference being that there are radio buttons to select the direction the list is printed in.