fixed initial form style
Now all forms, except for the one to select functionality, are set to have no display initially as opposed to being merely hidden. This is a subtle change but it is important for making the webpage look good. Before, when I was just toggling the form elements visibility, there would be massive amounts of white space between the top of the page and the form.