Added return statement to forwardPrint()
In changing the result of a print operation from a pop-up window to a webpage element, I need to have the print methods return a string. Since this string will be inserted in to an html document, I also had to change each "\r" to "<br>". I only modified the forwardPrint() method for testing purposes.