Fixed print form dead end
Now, when the user doesn't select a direction to print in a pop up warns them to select a print direction and they are asked again to select a print direction. Before, this was a dead end where they would only see the pop up followed by a blank page.